backend/stream-services/stage2-storage-fanout/requirements.txt
# Stage 2: Storage Fan-Out Service Dependencies
# Python service for multi-sink writes with FHIR transformations

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Kafka Integration
kafka-python==2.0.2
confluent-kafka==2.3.0

# Async Processing
asyncio==3.4.3
aiohttp==3.9.1
aiofiles==23.2.0

# Concurrent Processing (ThreadPoolExecutor)
concurrent-futures==3.1.1

# JSON Processing
orjson==3.9.10

# HTTP Client for API calls
httpx==0.25.2
requests==2.31.0

# Google Healthcare API (EXACT same as your PySpark implementation)
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Elasticsearch
elasticsearch==8.11.0
elasticsearch-async==6.2.0

# MongoDB
pymongo==4.6.0
motor==3.3.2  # Async MongoDB driver

# Redis (for caching and coordination)
redis==5.0.1
aioredis==2.0.1

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Configuration Management
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Date/Time Handling
python-dateutil==2.8.2
pytz==2023.3

# Validation and Data Processing
jsonschema==4.20.0
marshmallow==3.20.1

# Circuit Breaker Pattern
pybreaker==1.0.1

# Retry Logic
tenacity==8.2.3

# Health Checks
healthcheck==1.3.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx-mock==0.7.0

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Security
cryptography==41.0.8
